loanProject/requirements.txt
Django==4.2.7
django-extensions==3.2.1
django-allauth
plotly
scikit-learn==1.1.3
numpy==1.23.4
matplotlib==3.6.2
pandas==1.5.1
seaborn==0.12.1
tensorflow==2.8.0
keras==2.8.0